Set up Firebase Admin SDK using .env
